Task 6.Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
 */
 int main(void)
 { 
-(unsigned long)sizeof(d));
-(unsigned long)sizeof(a));
-(unsigned long)sizeof(b));
-(unsigned long)sizeof(c));
-(unsigned long)sizeof(f));
+sizeof(d));
+sizeof(a));
+sizeof(b));
+sizeof(c));
+sizeof(f));
 return (0)
 }
